PluginManager small refactoring
